requirements.txt
# requirements.txt - Project dependencies
python-telegram-bot==20.0
aiohttp==3.8.1
requests==2.26.0
